moved cart code to a separate component to simplify the header; made it so that the cart is cleared when an order is successfully placed
